WD: code-cleanup, reorganizing, added "useTime" hook, add: comments
