+++ client/src/components/clothesItemCreator/clothesIconCreator.tsx
