+++ client/src/components/clothesItemCreator/clothesItemCreator.tsx
