+++ client/src/pages/Faq/Faq.css
