+++ client/src/pages/Home/Home.tsx
@@ -13,8 +13,9 @@ import {
   IonCol,
   IonImg,
 } from "@ionic/react";
-import { Weather } from "components/WeatherAPI/Weather";
-import { Outfit } from "components/Outfit";
+
+import { Weather } from "components/WeatherAPI/renderWeather";
+import { Outfit } from "components/Outfit/Outfit";
 import { useOutfits } from "utils/api";
 import { useRedirect } from "utils/redirect";
 import { routes } from "utils/routes";
@@ -47,10 +48,10 @@ const Home: React.FC = () => {
         </IonCard>
         <IonGrid>
           <IonRow>
-            <IonCol size="5" style={{marginTop:"12%"}}>
+            <IonCol size="5" style={{ marginTop: "12%" }}>
               <IonImg class="helmet" src="assets/icon/custom/clothesIcons/iconHelmet3.png"></IonImg>
               <IonImg class="peep" src="assets/icon/custom/peeps/Robin.png"></IonImg>
-              <IonImg class="backpack" src="assets/icon/custom/backpack.png"></IonImg>              
+              <IonImg class="backpack" src="assets/icon/custom/backpack.png"></IonImg>
             </IonCol>
             <IonCol>
               <Outfit />
